refactored waveform plotting
